+Fixed even more warnings *Started working on the keyboard entry system

+++ Source/wl_inter.cpp
@@ -386,7 +386,7 @@ Write (int x, int y, const char *string)
                     break;
 
                 default:
-                    VWB_DrawPic (nx, ny, alpha[ch]);
+                    VWB_DrawPic (nx, ny, alpha[(uint8_t) ch]);
             }
             nx += 16;
         }
